Upgrade chain ID data type to u64 across the application
The data type of chain IDs has been changed from u32 to u64 across various components of the application, offering a wider range for chain IDs. The updates have been performed in 'src/error.rs', 'src/entities/currency.rs', 'src/entities/weth9.rs', 'src/addresses.rs', 'src/entities/base_currency.rs', 'src/entities/token.rs', 'src/entities/ether.rs', and a few other places. The library version has been updated in Cargo.lock and Cargo.toml to reflect the change.

+++ src/entities/token.rs
@@ -42,7 +42,7 @@ impl CurrencyTrait for Token {
 
 impl Token {
     pub fn new(
-        chain_id: u32,
+        chain_id: u64,
         address: String,
         decimals: u8,
         symbol: Option<String>,
